requirements.txt
fastapi~=0.111.0
uvicorn[standard]
sqlalchemy~=2.0.31
Cython~=3.0.10
PyMySQL~=1.1.1
pydantic~=2.8.2
alembic~=1.13.2
cryptography~=42.0.8
pyjwt~=2.8.0
passlib~=1.7.4
pydantic-settings~=2.3.4
